STL porting
git-svn-id: svn://svn.savannah.gnu.org/myserver/trunk@1355 c4c922ea-a70d-0410-937d-ec1daf91936c

+++ myserverweb/source/processes.cpp
@@ -75,7 +75,7 @@ int Process::execHiddenProcess(StartProcInfo *spi,u_long timeout)
                       spi->envString, cwd, &si, &pi);
 	if(!ret)
 		return (-1);
-  pid = pi.hProcess;
+  pid = (u_long)pi.hProcess;
 	/*!
 	*Wait until the process stops its execution.
 	*/
